Makes 2 asteroids with random velocities
Don't know how much of it is running them all on one computer, but it
takes a bit for the ships to get synced up with the asteroids in.

+++ include/GameState_NetworkPlay.h
@@ -30,5 +30,5 @@ private:
 	ProtoConnection * m_gameServer;
 	int m_netID;
 	int m_send;
-	int m_lastRecv;
+	float m_lastRecv;
 };
